Shuffling around builtin types/type_defs/etc to prepare for list impl.
