+++ iree/hal/vmla/BUILD
@@ -159,7 +159,6 @@ cc_library(
         "//iree/base:tracing",
         "//iree/vm",
         "//iree/vm:module_abi_cc",
-        "//iree/vm:types",
         "@com_google_absl//absl/types:span",
     ],
 )
