+++ iree/hal/vmla/vmla_module.h
@@ -28,7 +28,6 @@
 #include "iree/base/status.h"
 #include "iree/vm/api.h"
 #include "iree/vm/module_abi_cc.h"
-#include "iree/vm/types.h"
 
 namespace iree {
 namespace hal {
