+++ iree/modules/strings/BUILD
@@ -28,7 +28,6 @@ cc_library(
         "//iree/vm:ref",
         "//iree/vm:ref_cc",
         "//iree/vm:stack",
-        "//iree/vm:types",
         "@com_google_absl//absl/container:inlined_vector",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/types:span",
@@ -55,7 +54,6 @@ cc_test(
         "//iree/vm:ref",
         "//iree/vm:ref_cc",
         "//iree/vm:stack",
-        "//iree/vm:types",
         "@com_google_absl//absl/container:inlined_vector",
         "@com_google_absl//absl/strings",
         "@com_google_benchmark//:benchmark",
