+++ iree/vm/CMakeLists.txt
@@ -14,12 +14,26 @@
 
 iree_add_all_subdirs()
 
+iree_cc_library(
+  NAME
+    builtin_types
+  HDRS
+    "builtin_types.h"
+  SRCS
+    "builtin_types.c"
+  DEPS
+    ::ref
+    iree::base::api
+  PUBLIC
+)
+
 iree_cc_test(
   NAME
     bytecode_dispatch_test
   SRCS
     "bytecode_dispatch_test.cc"
   DEPS
+    ::builtin_types
     ::bytecode_module
     ::context
     ::instance
@@ -42,11 +56,13 @@ iree_cc_library(
     "bytecode_module_impl.h"
     "bytecode_op_table.h"
   DEPS
+    ::builtin_types
     ::module
     ::ref
     ::stack
-    ::types
+    ::type_def
     ::value
+    absl::strings
     flatbuffers
     iree::base::alignment
     iree::base::api
@@ -132,7 +148,7 @@ iree_cc_library(
   SRCS
     "instance.c"
   DEPS
-    ::types
+    ::builtin_types
     iree::base::api
     iree::base::atomics
   PUBLIC
@@ -175,11 +191,11 @@ iree_cc_library(
     "module_abi_cc.h"
     "module_abi_packing.h"
   DEPS
+    ::builtin_types
     ::module
     ::ref
     ::ref_cc
     ::stack
-    ::types
     absl::span
     absl::strings
     iree::base::api
@@ -221,7 +237,6 @@ iree_cc_library(
     "ref_cc.h"
   DEPS
     ::ref
-    ::types
     absl::core_headers
     iree::base::api
   PUBLIC
@@ -259,14 +274,12 @@ iree_cc_test(
 
 iree_cc_library(
   NAME
-    types
+    type_def
   HDRS
-    "types.h"
-  SRCS
-    "types.c"
+    "type_def.h"
   DEPS
     ::ref
-    iree::base::api
+    ::value
   PUBLIC
 )
 
@@ -297,13 +310,14 @@ iree_cc_library(
   HDRS
     "api.h"
   DEPS
+    ::builtin_types
     ::context
     ::instance
     ::invocation
     ::module
     ::ref
     ::stack
-    ::types
+    ::type_def
     ::value
     ::variant_list
     iree::base::api
