+++ iree/vm/api.h
@@ -16,13 +16,13 @@
 #define IREE_VM_API_H_
 
 #include "iree/base/api.h"
+#include "iree/vm/builtin_types.h"
 #include "iree/vm/context.h"
 #include "iree/vm/instance.h"
 #include "iree/vm/invocation.h"
 #include "iree/vm/module.h"
 #include "iree/vm/ref.h"
 #include "iree/vm/stack.h"
-#include "iree/vm/types.h"
 #include "iree/vm/value.h"
 #include "iree/vm/variant_list.h"
