+++ iree/vm/bytecode_module.cc
@@ -44,8 +44,14 @@ static iree_vm_type_def_t iree_vm_bytecode_module_resolve_type(
   auto full_name = iree::WrapString(type_def->full_name());
   iree_vm_type_def_t result;
   memset(&result, 0, sizeof(result));
-  if (full_name == "i32") {
+  if (full_name == "i8") {
+    result.value_type = IREE_VM_VALUE_TYPE_I8;
+  } else if (full_name == "i16") {
+    result.value_type = IREE_VM_VALUE_TYPE_I16;
+  } else if (full_name == "i32") {
     result.value_type = IREE_VM_VALUE_TYPE_I32;
+  } else if (full_name == "i64") {
+    result.value_type = IREE_VM_VALUE_TYPE_I64;
   } else if (!full_name.empty() && full_name[0] == '!') {
     full_name.remove_prefix(1);
     const iree_vm_ref_type_descriptor_t* type_descriptor =
